requirements.txt
openai-agents>=0.11.1
openai>=2.25.0
fastapi>=0.110
uvicorn>=0.29
streamlit>=1.35
python-dotenv>=1.0
requests>=2.31
